Update cast error and fix settings

+++ bonus/src/floor_ceiling_render/render_data.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   render_data.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bgoron <bgoron@student.42.fr>              +#+  +:+       +#+        */
+/*   By: asuc <asuc@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/22 02:42:52 by asuc              #+#    #+#             */
-/*   Updated: 2024/06/27 15:31:01 by bgoron           ###   ########.fr       */
+/*   Updated: 2024/07/02 05:58:39 by asuc             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@ inline void	init_pixel_data(t_pixel_data *pd, float floor_x,
 	const int	shift = 20;
 	const int	fixed_1 = 1 << shift;
 
-	pd->floor_xi = (int64_t)(floor_x * fixed_1);
-	pd->floor_yi = (int64_t)(floor_y * fixed_1);
-	pd->floor_step_xi = (int64_t)(floor_step.x * fixed_1);
-	pd->floor_step_yi = (int64_t)(floor_step.y * fixed_1);
+	pd->floor_xi = (int64_t)(floor_x * (float)fixed_1);
+	pd->floor_yi = (int64_t)(floor_y * (float)fixed_1);
+	pd->floor_step_xi = (int64_t)(floor_step.x * (float)fixed_1);
+	pd->floor_step_yi = (int64_t)(floor_step.y * (float)fixed_1);
 }
 
 inline void	init_render_data(t_render_data *rd, t_data *data)
@@ -31,8 +31,8 @@ inline void	init_render_data(t_render_data *rd, t_data *data)
 	rd->ray_dir_y0 = data->player.dir.y - data->player.plane.y;
 	rd->ray_dir_x1 = data->player.dir.x + data->player.plane.x;
 	rd->ray_dir_y1 = data->player.dir.y + data->player.plane.y;
-	rd->center_line = WIN_HEIGHT / 2.0;
-	rd->pos_z = 0.5 * WIN_HEIGHT;
+	rd->center_line = WIN_HEIGHT / 2.0f;
+	rd->pos_z = 0.5f * WIN_HEIGHT;
 }
 
 inline void	calculate_floor_step(t_render_data *rd, float row_distance,
