+++ bonus/src/minimap/minimap_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minimap_utils.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bgoron <bgoron@student.42.fr>              +#+  +:+       +#+        */
+/*   By: asuc <asuc@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/01 15:17:11 by bgoron            #+#    #+#             */
-/*   Updated: 2024/07/01 15:19:25 by bgoron           ###   ########.fr       */
+/*   Updated: 2024/07/02 05:50:35 by asuc             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,9 +45,9 @@ void	set_minimap_color(t_data *d, t_vec2f map, int *color)
 	if (map.x >= 0 && map.x < d->map.cols && map.y >= 0 && map.y < d->map.rows)
 	{
 		if (d->map.map[(int)map.y][(int)map.x] == '1')
-			*color = 0xFFFF0000;
+			*color = (int)(uintptr_t)0xFFFF0000;
 		if (d->map.map[(int)map.y][(int)map.x] == 'D')
-			*color = 0xFF0000FF;
+			*color = (int)(uintptr_t)0xFF0000FF;
 	}
 }
 
@@ -59,17 +59,17 @@ void	set_minimap_pixel(t_data *d, int *mask, t_vec2i pos, t_vec2f p_map)
 	int		color;
 	int		size;
 
-	size = (WIN_HEIGHT * 0.56 / 3) * d->settings.minimap_scale;
-	cos_sin.x = cos(atan2(d->player.dir.y, d->player.dir.x));
-	cos_sin.y = sin(atan2(d->player.dir.y, d->player.dir.x));
-	ratio = (WIN_HEIGHT * 0.28 / 3) * d->settings.minimap_scale;
+	size = (int)((WIN_HEIGHT * 0.56f / 3.0f) * d->settings.minimap_scale);
+	cos_sin.x = cosf(atan2f(d->player.dir.y, d->player.dir.x));
+	cos_sin.y = sinf(atan2f(d->player.dir.y, d->player.dir.x));
+	ratio = (WIN_HEIGHT * 0.28f / 3.0f) * d->settings.minimap_scale;
 	if (mask[pos.y * size + pos.x])
 	{
-		map.x = p_map.x + (cos_sin.x * (-(pos.y - ratio)) - cos_sin.y \
-		* (pos.x - ratio)) / MINIMAP_SCALE * d->settings.minimap_scale;
-		map.y = p_map.y + (cos_sin.y * (-(pos.y - ratio)) + cos_sin.x \
-		* (pos.x - ratio)) / MINIMAP_SCALE * d->settings.minimap_scale;
-		color = 0xFF808080;
+		map.x = p_map.x + (cos_sin.x * (-((float)pos.y - ratio)) - cos_sin.y \
+		* ((float)pos.x - ratio)) / MINIMAP_SCALE * d->settings.minimap_scale;
+		map.y = p_map.y + (cos_sin.y * (-((float)pos.y - ratio)) + cos_sin.x \
+		* ((float)pos.x - ratio)) / MINIMAP_SCALE * d->settings.minimap_scale;
+		color = (int)(uintptr_t)0xFF808080;
 		set_minimap_color(d, map, &color);
 		mlx_set_image_pixel(d->graphics.mlx, \
 		d->graphics.img_mini_map, pos.x, pos.y, color);
