+++ bonus/src/movement/keyboard_movement.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   keyboard_movement.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bgoron <bgoron@student.42.fr>              +#+  +:+       +#+        */
+/*   By: asuc <asuc@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/14 16:54:33 by asuc              #+#    #+#             */
-/*   Updated: 2024/06/25 14:51:45 by bgoron           ###   ########.fr       */
+/*   Updated: 2024/07/02 01:11:58 by asuc             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,9 +67,9 @@ void	rotate(t_player *player, float angle)
 
 	old_plane_x = player->plane.x;
 	old_dir_x = player->dir.x;
-	player->dir.x = player->dir.x * cos(angle) - player->dir.y * sin(angle);
-	player->dir.y = old_dir_x * sin(angle) + player->dir.y * cos(angle);
-	player->plane.x = player->plane.x * cos(angle) - player->plane.y
-		* sin(angle);
-	player->plane.y = old_plane_x * sin(angle) + player->plane.y * cos(angle);
+	player->dir.x = player->dir.x * cosf(angle) - player->dir.y * sinf(angle);
+	player->dir.y = old_dir_x * sinf(angle) + player->dir.y * cosf(angle);
+	player->plane.x = player->plane.x * cosf(angle) - player->plane.y
+		* sinf(angle);
+	player->plane.y = old_plane_x * sinf(angle) + player->plane.y * cosf(angle);
 }
