+++ bonus/src/parsing/check_map.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   check_map.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bgoron <bgoron@student.42.fr>              +#+  +:+       +#+        */
+/*   By: asuc <asuc@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/13 17:11:56 by bgoron            #+#    #+#             */
-/*   Updated: 2024/06/28 14:30:21 by bgoron           ###   ########.fr       */
+/*   Updated: 2024/07/02 04:36:31 by asuc             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,28 +36,28 @@ static void	set_player_dir_and_plane(char **map, int i, int j, t_player *player)
 {
 	if (map[i][j] == 'N')
 	{
-		player->dir = (t_vec2f){0, -1};
-		player->plane = (t_vec2f){0.66, 0};
+		player->dir = (t_vec2f){0.0f, -1.0f};
+		player->plane = (t_vec2f){0.66f, 0.0f};
 	}
 	else if (map[i][j] == 'S')
 	{
-		player->dir = (t_vec2f){0, 1};
-		player->plane = (t_vec2f){-0.66, 0};
+		player->dir = (t_vec2f){0.0f, 1.0f};
+		player->plane = (t_vec2f){-0.66f, 0.0f};
 	}
 	else if (map[i][j] == 'E')
 	{
-		player->dir = (t_vec2f){1, 0};
-		player->plane = (t_vec2f){0, 0.66};
+		player->dir = (t_vec2f){1.0f, 0.0f};
+		player->plane = (t_vec2f){0.0f, 0.66f};
 	}
 	else if (map[i][j] == 'W')
 	{
-		player->dir = (t_vec2f){-1, 0};
-		player->plane = (t_vec2f){0, -0.66};
+		player->dir = (t_vec2f){-1.0f, 0.0f};
+		player->plane = (t_vec2f){0.0f, -0.66f};
 	}
 	map[i][j] = '0';
-	player->pos = (t_vec2f){j + 0.5, i + 0.5};
-	player->move_speed = 0.05;
-	player->rot_speed = 0.05;
+	player->pos = (t_vec2f){(float)j + 0.5f, (float)i + 0.5f};
+	player->move_speed = 0.05f;
+	player->rot_speed = 0.05f;
 }
 
 int	check_player(char **map, t_player *player)
